requirements.txt
# Core dependencies
setuptools>=65.5.1
wheel>=0.38.4

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
dnspython==2.4.2

# Data processing and visualization
numpy>=1.24.3
matplotlib>=3.7.1
plotly>=5.18.0
seaborn>=0.12.2
wordcloud>=1.9.2

# PDF generation
reportlab>=4.0.4

# LLM support
llama-cpp-python>=0.2.11

# Utilities
python-dotenv>=1.0.0